+++ src/api/services/util.js
@@ -10,6 +10,7 @@ const UPORT_PUSH_TOKEN_HEADER = 'Uport-Push-Token' // deprecated: use Authorizat
 
 const ERROR_CODES = {
   CANNOT_REGISTER_TOO_SOON: 'CANNOT_REGISTER_TOO_SOON',
+  INVALID_AUDIENCE: 'INVALID_AUDIENCE', // copied from ./vc/index.js, not 'import'ed because we've been running this in raw JS sql-by-hand scripts
   JWT_VERIFY_FAILED: 'JWT_VERIFY_FAILED', // copied from ./vc/index.js, not 'import'ed because we've been running this in raw JS sql-by-hand scripts
   OVER_CLAIM_LIMIT: 'OVER_CLAIM_LIMIT',
   OVER_REGISTRATION_LIMIT: 'OVER_REGISTRATION_LIMIT',